Added some simulated Data files.
Using Machine  learning we are able to filter background efficiently.
Also the lowZ material is nicely segregated using Random Forest.
But when we move towards medium to High Z material then classification
score start reducing, So its difficult to segregate Pb from U

The reason may be, because with highZ  we get broader scattering gaussian,
which overlaps with gaussians of others, and hence less classification score.